gse-demo: use jpackager to create a native image
Tested on linux only.
This shouldn't work on windows because the docs say
"
For Windows, you have to unzip the bundle in the same directory as the JDK,
as the packager wrapper expect to find the java binary at the same level.
"

Maybe it will work in the future with a new release from
http://download2.gluonhq.com/jpackager/11/jdk.packager-windows.zip

In the meantime, users should be able to move the downloaded jpackager
to their jdk themselves and invoke maven with the correct path to the packager
with something like this (untested):
mvn -Djpackager.path=C:\\Path\\To\\JDK\\bin\\jpackager